+++ src/constants.ts
@@ -1,27 +1,18 @@
 import { OpenCVBuilder } from "./OpenCVBuilder.js"
 
-const opencvModules = [
-  'core',
-  'highgui',
-  'imgcodecs',
-  'imgproc',
-  'features2d',
-  'calib3d',
-  'photo',
-  'objdetect',
-  'ml',
-  'video',
-  'videoio',
-  'videostab',
-  'dnn',
-  'face',
-  'text',
-  'tracking',
-  'xfeatures2d',
-  'ximgproc'
-] as const;
+const opencvModules = [ 'apps', 'aruco', 'bgsegm', 'bioinspired', 'calib3d', 'ccalib',
+'core', 'datasets', 'dnn', 'dnn_objdetect', 'dpm', 'features2d', 'flann', 'fuzzy',
+'gapi', 'hfs', 'highgui', 'img_hash', 'imgcodecs', 'imgproc', 'java_bindings_generator',
+'js', 'js_bindings_generator', 'line_descriptor', 'ml', 'objc_bindings_generator',
+'objdetect', 'optflow', 'phase_unwrapping', 'photo', 'python3', 'python_bindings_generator',
+'python_tests', 'reg', 'rgbd', 'saliency', 'shape', 'stereo', 'stitching', 'structured_light',
+'superres', 'surface_matching', 'ts', 'video', 'videoio', 'wechat_qrcode', 'world',
+'xobjdetect', 'xphoto'] as const;
+export type OpencvModulesType = typeof opencvModules[number] | 'world';
 
-export type opencvModulesType = typeof opencvModules[number] | 'world';
+export const enabledModules = new Set<OpencvModulesType>(
+  [ 'calib3d', 'core', 'dnn', 'features2d', 'flann', 'gapi', 'highgui', 'imgcodecs', 'imgproc',
+  'ml', 'objdetect', 'photo', 'python_tests', 'video', 'videoio']);
 
 export class Constant {
   constructor(private readonly builder: OpenCVBuilder) {
@@ -29,9 +20,10 @@ export class Constant {
 
   opencvRepoUrl = 'https://github.com/opencv/opencv.git'
   // opencvRepoUrl = 'c:/cache/opencv'
+
   opencvContribRepoUrl = 'https://github.com/opencv/opencv_contrib.git'
   // opencvContribRepoUrl = 'c:/cache/opencv_contrib'
-  
+
   opencvModules = opencvModules;
 
   cmakeVsCompilers: { [version: string]: string } = {
@@ -43,14 +35,19 @@ export class Constant {
     '16': 'Visual Studio 16 2019',
     '17': 'Visual Studio 17 2022',
   }
+  cmakeArchs: { [arch: string]: string } = { 'x64': ' Win64', 'ia32': '', 'arm': ' ARM' }
 
-  cmakeArchs: { [arch: string]: string } = {
-    'x64': ' Win64',
-    'ia32': '',
-    'arm': ' ARM'
+  defaultCmakeBuildFlags(): string[] {
+    const out: string[] = [];
+    for (const mod of opencvModules) {
+      let arg =`-DBUILD_opencv_${mod}=`;
+      arg += enabledModules.has(mod) ? 'ON' : 'OFF';
+      out.push(arg);
+    }
+    return out;
   }
 
-  defaultCmakeFlags = () => {
+  public defaultCmakeFlags = () => {
     return [
       `-DCMAKE_INSTALL_PREFIX=${this.builder.env.opencvBuild}`,
       '-DCMAKE_BUILD_TYPE=Release',
@@ -59,39 +56,15 @@ export class Constant {
       '-DBUILD_TESTS=OFF',
       '-DBUILD_PERF_TESTS=OFF',
       '-DBUILD_JAVA=OFF',
+      '-DBUILD_ZLIB=OFF', // https://github.com/opencv/opencv/issues/21389
       '-DCUDA_NVCC_FLAGS=--expt-relaxed-constexpr',
-      '-DBUILD_opencv_apps=OFF',
-      '-DBUILD_opencv_aruco=OFF',
-      '-DBUILD_opencv_bgsegm=OFF',
-      '-DBUILD_opencv_bioinspired=OFF',
-      '-DBUILD_opencv_ccalib=OFF',
-      '-DBUILD_opencv_datasets=OFF',
-      '-DBUILD_opencv_dnn_objdetect=OFF',
-      '-DBUILD_opencv_dpm=OFF',
-      '-DBUILD_opencv_fuzzy=OFF',
-      '-DBUILD_opencv_hfs=OFF',
-      '-DBUILD_opencv_java_bindings_generator=OFF',
-      '-DBUILD_opencv_js=OFF',
-      '-DBUILD_opencv_img_hash=OFF',
-      '-DBUILD_opencv_line_descriptor=OFF',
-      '-DBUILD_opencv_optflow=OFF',
-      '-DBUILD_opencv_phase_unwrapping=OFF',
-      '-DBUILD_opencv_python3=OFF',
-      '-DBUILD_opencv_python_bindings_generator=OFF',
-      '-DBUILD_opencv_reg=OFF',
-      '-DBUILD_opencv_rgbd=OFF',
-      '-DBUILD_opencv_saliency=OFF',
-      '-DBUILD_opencv_shape=OFF',
-      '-DBUILD_opencv_stereo=OFF',
-      '-DBUILD_opencv_stitching=OFF',
-      '-DBUILD_opencv_structured_light=OFF',
-      '-DBUILD_opencv_superres=OFF',
-      '-DBUILD_opencv_surface_matching=OFF',
-      '-DBUILD_opencv_ts=OFF',
-      '-DBUILD_opencv_xobjdetect=OFF',
-      '-DBUILD_opencv_xphoto=OFF',
-      '-DBUILD_opencv_wechat_qrcode=OFF', // since 10/05/2022
-      '-DWITH_VTK=OFF'
+      '-DWITH_VTK=OFF',
+      ...this.defaultCmakeBuildFlags()
     ];
   }
+  // if version < 4.5.6 ffmpeg 5 not compatible
+  // https://stackoverflow.com/questions/71070080/building-opencv-from-source-in-mac-m1
+  // brew install ffmpeg@4
+  // brew unlink ffmpeg
+  // brew link ffmpeg@4
 }